\name{zoo-class}
\Rdversion{1.1}
\docType{class}
\alias{zoo-class}

\title{Class \code{"zoo"}}
\description{
  This is the class \code{zoo} from package zoo, made available for use
  in S4 classes.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
  No methods defined with class "zoo" in the signature.
}
\author{
Boris Orlowsky <boris@climate-babel.org>
}
\examples{
showClass("zoo")
}
\keyword{classes}
